bugfix and cleanup, rework of footer and single pages
fixed a bug where loading the footer object oriented caused a crash of the webpage due to not checking if child items are correctly linked to parents.
deleted footer.php due to resulting code redundancy and created a function in navigation to load the footer properly.
removed footer loading functions from dbmanager and merging the functionality into getNavigationItems by adding a parameter ($table).